requirements.txt
django
pywebview

django-jazzmin
